pyproject.toml
[project]
name = "gaze-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "coloredlogs>=15.0.1",
    "dlib-bin>=19.24.6",
    "greenlet>=3.1.1",
    "opencv-contrib-python>=4.10.0.84",
    "opencv-python>=4.10.0.84",
    "pillow>=11.1.0",
    "ptext>=1.1.0",
    "pygame>=2.6.1",
    "pytelegrambotapi>=4.26.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "icecream>=2.1.4",
    "wat>=0.5.0",
]
